refactor: :recycle: rename CounselCardService package and update imports
